Refacto WS tests
